change type of  cost, some refactoring changes

+++ src/domain/domain.go
@@ -9,7 +9,7 @@ type Route struct {
 	ID        int       `json:"id"`
 	Points    Points    `json:"points"`
 	Start     time.Time `json:"start_time"`
-	Cost      float32   `json:"cost"`
+	Cost      int       `json:"cost"`
 	FreeSeats int       `json:"freeseats"`
 	AllSeats  int       `json:"allseats"`
 }
